requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
distro==1.8.0
Django==4.2.6
et-xmlfile==1.1.0
idna==3.4
JPype1==1.4.1
numpy==1.26.0
openpyxl==3.1.2
packaging==23.2
pandas==2.1.1
PyMySQL==1.1.0
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tabula-py==2.8.2
tzdata==2023.3
urllib3==2.0.6
xlrd==2.0.1